pyproject.toml
[tool.poetry]
name = "play-tennis-ui"
version = "0.1.0"
description = "Create web app with streamlit for developers for interactive inference"
authors = ["roniabrahamTA <roni.abraham@tigeranalytics.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.33.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
show-fixes = true
line-length = 88
target-version = "py310"
namespace-packages = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 88
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "C4",
    "C90",
    "E",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "LOG",
    "PT",
    "N",
    "Q",
    "RUF",
    "SIM",
    "SLF",
    "TCH",
    "TID",
    "TD",
    "UP",
    "YTT",
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "docs",
    ".pytest_cache",
    "docs/source/conf",
]
ignore = [
    "D100",
    "D101",
    "D102",
    "D104",
    "D107",
    "D417",
    "D206",
    "D300",
    "E402",
    "E111",
    "E114",
    "E117",
    "E731",
    "F403",
    "ISC001",
    "ISC002",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "N812",
    "TD002",
    "TD003",
    "W191",
]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.lint.flake8-pytest-style]
parametrize-names-type = "csv"

[tool.ruff.lint.flake8-type-checking]
strict = false

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.pep8-naming]
extend-ignore-names = ["X", "Y", "Z"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
combine-as-imports = true
detect-same-package = true
extra-standard-library = []
force-sort-within-sections = true
from-first = false
known-first-party = []
known-local-folder = []
known-third-party = []